requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
junit-xml==1.9
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
prometheus_client==0.20.0
pytest==8.1.1
six==1.16.0
splitio-client==10.5.1
tomli==2.0.1
Werkzeug==3.0.3
